chore(cli): Refactor pkg/cli envoy helpers.
This refactors pkg/cli envoy helpers and addresses
some of the PR comments mentioned in #3881.
